split all
