+++ src/puppet-mixins/contact-self.ts
@@ -1,9 +1,7 @@
 import {
-  log,
+  log, PRE,
 } from '../config.js'
 import type PuppetWhatsapp from '../puppet-whatsapp'
-
-const PRE = 'PuppetWhatsApp'
 
 export async function  contactSelfQRCode (this:PuppetWhatsapp): Promise<string> {
   log.verbose(PRE, 'contactSelfQRCode()')
@@ -13,4 +11,9 @@ export async function  contactSelfQRCode (this:PuppetWhatsapp): Promise<string> 
 export async function contactSelfName (this:PuppetWhatsapp, name: string): Promise<void> {
   log.verbose(PRE, 'contactSelfName(%s)', name)
   await this.getManager()!.setNickname(name)
+}
+
+export async function contactSelfSignature (this:PuppetWhatsapp, signature: string): Promise<void> {
+  log.verbose(PRE, 'contactSelfSignature(%s)', signature)
+  await this.getWhatsapp()!.setStatus(signature)
 }
